+++ NEWS.md
@@ -4,4 +4,6 @@
 
 # Version 1.7.7
 - New function 'predict.indicators.cv' for cross validation of predictive value
-- New option 'min.order' in function 'indicators
+- New option 'min.order' in function 'indicators'
+- New algorithm for 'combinespecies', similar to the one used for 'indicators'
+- New option 'min.order' in function 'combinespecies'